+++ com.unity.cinemachine/Editor/Editors/CinemachineFreeLookEditor.cs
@@ -60,6 +60,17 @@ namespace Cinemachine
 
         public override void OnInspectorGUI()
         {
+            Target.GetRig(0); // force an update of the rig cache
+            if (!Target.RigsAreCreated)
+            {
+                EditorGUILayout.HelpBox(
+                    "It's not possible to add this component to a prefab instance.  "
+                    + "Instead, you can open the Prefab in Prefab Mode or unpack "
+                    + "the Prefab instance to remove the Prefab connection.",
+                    MessageType.Error);
+                return;
+            }
+            
             Target.m_XAxis.ValueRangeLocked
                 = (Target.m_BindingMode == CinemachineTransposer.BindingMode.SimpleFollowWithWorldUp);
 
@@ -221,27 +232,54 @@ namespace Cinemachine
             {
                 CinemachineFreeLook.CreateRigOverride
                     = (CinemachineFreeLook vcam, string name, CinemachineVirtualCamera copyFrom) =>
+                {
+                    // Recycle the game object if it exists
+                    GameObject go = null;
+                    foreach (Transform child in vcam.transform)
                     {
-                        // Create a new rig with default components
-                        GameObject go = ObjectFactory.CreateGameObject(name);
-                        Undo.RegisterCreatedObjectUndo(go, "created rig");
-                        Undo.SetTransformParent(go.transform, vcam.transform, "parenting rig");
-                        CinemachineVirtualCamera rig = Undo.AddComponent<CinemachineVirtualCamera>(go);
-                        Undo.RecordObject(rig, "creating rig");
-                        if (copyFrom != null)
-                            ReflectionHelpers.CopyFields(copyFrom, rig);
-                        else
+                        if (child.gameObject.name == name)
                         {
-                            go = rig.GetComponentOwner().gameObject;
-                            Undo.RecordObject(Undo.AddComponent<CinemachineOrbitalTransposer>(go), "creating rig");
-                            Undo.RecordObject(Undo.AddComponent<CinemachineComposer>(go), "creating rig");
+                            go = child.gameObject;
+                            break;
                         }
-                        return rig;
-                    };
-                CinemachineFreeLook.DestroyRigOverride = (GameObject rig) =>
+                    }
+
+                    CinemachineVirtualCamera rig = null;
+                    if (go == null)
                     {
-                        Undo.DestroyObjectImmediate(rig);
-                    };
+                        // Create a new rig - can't do it if prefab instance
+                        if (PrefabUtility.IsPartOfAnyPrefab(vcam.gameObject))
+                            return null;
+                        go =  ObjectFactory.CreateGameObject(name);
+                        Undo.RegisterCreatedObjectUndo(go, "created rig");
+                        Undo.SetTransformParent(go.transform, vcam.transform, "parenting pipeline");
+                    }
+
+                    // Create a new rig with default components
+                    rig = Undo.AddComponent<CinemachineVirtualCamera>(go);
+                    rig.CreatePipeline(copyFrom);
+                    if (copyFrom != null)
+                        ReflectionHelpers.CopyFields(copyFrom, rig);
+                    else
+                    {
+                        // Defaults
+                        go = rig.GetComponentOwner().gameObject;
+                        Undo.AddComponent<CinemachineOrbitalTransposer>(go);
+                        Undo.AddComponent<CinemachineComposer>(go);
+                        rig.InvalidateComponentPipeline();
+                    }
+                    return rig;
+                };
+
+                CinemachineFreeLook.DestroyRigOverride = (GameObject rig) =>
+                {
+                    var vcam = rig.GetComponent<CinemachineVirtualCamera>();
+                    if (vcam != null)
+                    {
+                        vcam.DestroyPipeline();
+                        Undo.DestroyObjectImmediate(vcam);
+                    }
+                };
             }
         }