+++ com.unity.cinemachine/Runtime/Behaviours/CinemachineVirtualCamera.cs
@@ -214,16 +214,8 @@ namespace Cinemachine
 
         void Reset()
         {
-#if UNITY_EDITOR
-            if (UnityEditor.PrefabUtility.GetPrefabInstanceStatus(gameObject)
-                != UnityEditor.PrefabInstanceStatus.NotAPrefab)
-            {
-                Debug.Log("You cannot reset a prefab instance.  "
-                    + "First disconnect this instance from the prefab, or enter Prefab Edit mode");
-                return;
-            }
-#endif
             DestroyPipeline();
+            UpdateComponentPipeline();
         }
 
         /// <summary>
@@ -255,30 +247,34 @@ namespace Cinemachine
         public delegate void DestroyPipelineDelegate(GameObject pipeline);
 
         /// <summary>Destroy any existing pipeline container.</summary>
-        private void DestroyPipeline()
+        internal void DestroyPipeline()
         {
             List<Transform> oldPipeline = new List<Transform>();
             foreach (Transform child in transform)
                 if (child.GetComponent<CinemachinePipeline>() != null)
                     oldPipeline.Add(child);
 
-            if (!RuntimeUtility.IsPrefab(gameObject))
+            foreach (Transform child in oldPipeline)
             {
-                foreach (Transform child in oldPipeline)
+                if (DestroyPipelineOverride != null)
+                    DestroyPipelineOverride(child.gameObject);
+                else
                 {
-                    if (DestroyPipelineOverride != null)
-                        DestroyPipelineOverride(child.gameObject);
-                    else
+                    var oldStuff = child.GetComponents<CinemachineComponentBase>();
+                    foreach (var c in oldStuff)
+                        Destroy(c);
+                    if (!RuntimeUtility.IsPrefab(gameObject))
                         Destroy(child.gameObject);
                 }
-                m_ComponentOwner = null;
             }
+            m_ComponentOwner = null;
+            InvalidateComponentPipeline();
             PreviousStateIsValid = false;
         }
 
         /// <summary>Create a default pipeline container.
         /// Note: copyFrom only supported in Editor, not build</summary>
-        private Transform CreatePipeline(CinemachineVirtualCamera copyFrom)
+        internal Transform CreatePipeline(CinemachineVirtualCamera copyFrom)
         {
             CinemachineComponentBase[] components = null;
             if (copyFrom != null)
@@ -290,9 +286,9 @@ namespace Cinemachine
             Transform newPipeline = null;
             if (CreatePipelineOverride != null)
                 newPipeline = CreatePipelineOverride(this, PipelineName, components);
-            else
+            else if (!RuntimeUtility.IsPrefab(gameObject))
             {
-                GameObject go =  new GameObject(PipelineName);
+                GameObject go = new GameObject(PipelineName);
                 go.transform.parent = transform;
                 go.AddComponent<CinemachinePipeline>();
                 newPipeline = go.transform;
@@ -394,22 +390,26 @@ namespace Cinemachine
         CameraState m_State = CameraState.Default; // Current state this frame
 
         CinemachineComponentBase[] m_ComponentPipeline = null;
-        [SerializeField][HideInInspector] private Transform m_ComponentOwner = null;   // serialized to handle copy/paste
+
+        // Serialized only to implement copy/paste of CM subcomponents.
+        // Note however that this strategy has its limitations: the CM pipeline Components
+        // won't be pasted onto a prefab asset outside the scene unless the prefab
+        // is opened in Prefab edit mode.
+        [SerializeField][HideInInspector] private Transform m_ComponentOwner = null;
+
         void UpdateComponentPipeline()
         {
-            bool isPrefab = RuntimeUtility.IsPrefab(gameObject);
 #if UNITY_EDITOR
             // Did we just get copy/pasted?
             if (m_ComponentOwner != null && m_ComponentOwner.parent != transform)
             {
-                if (!isPrefab) // can't paste to a prefab
-                {
-                    CinemachineVirtualCamera copyFrom = (m_ComponentOwner.parent != null)
-                        ? m_ComponentOwner.parent.gameObject.GetComponent<CinemachineVirtualCamera>() : null;
-                    DestroyPipeline();
-                    m_ComponentOwner = CreatePipeline(copyFrom);
-                }
+                CinemachineVirtualCamera copyFrom = (m_ComponentOwner.parent != null)
+                    ? m_ComponentOwner.parent.gameObject.GetComponent<CinemachineVirtualCamera>() : null;
+                DestroyPipeline();
+                CreatePipeline(copyFrom);
+                m_ComponentOwner = null;
             }
+            // Make sure the pipeline stays hidden, even through prefab
             if (m_ComponentOwner != null)
                 SetFlagsForHiddenChild(m_ComponentOwner.gameObject);
 #endif
@@ -423,21 +423,19 @@ namespace Cinemachine
             {
                 if (child.GetComponent<CinemachinePipeline>() != null)
                 {
-                    m_ComponentOwner = child;
                     CinemachineComponentBase[] components = child.GetComponents<CinemachineComponentBase>();
                     foreach (CinemachineComponentBase c in components)
                         if (c.enabled)
                             list.Add(c);
+                    m_ComponentOwner = child;
+                    break;
                 }
             }
 
             // Make sure we have a pipeline owner
-            if (m_ComponentOwner == null && !isPrefab)
+            if (m_ComponentOwner == null)
                 m_ComponentOwner = CreatePipeline(null);
 
-            // Make sure the pipeline stays hidden, even through prefab
-            if (m_ComponentOwner != null)
-                SetFlagsForHiddenChild(m_ComponentOwner.gameObject);
             if (m_ComponentOwner != null && m_ComponentOwner.gameObject != null)
             {
                 // Sort the pipeline